7_DocumentIntelligence/Layout/requirements.txt
azure-core
azure-ai-documentintelligence